+++ apps/onboarding-functions/src/main/java/it/pagopa/selfcare/onboarding/functions/utils/ActivityName.java
@@ -1,0 +1,20 @@
+package it.pagopa.selfcare.onboarding.functions.utils;
+
+public class ActivityName {
+
+    public static final String SAVE_TOKEN_WITH_CONTRACT_ACTIVITY_NAME = "SaveTokenWithContract";
+    public static final String BUILD_CONTRACT_ACTIVITY_NAME = "BuildContract";
+    public static final String SEND_MAIL_REGISTRATION_WITH_CONTRACT_ACTIVITY = "SendMailRegistrationWithContract";
+
+    public static final String SEND_MAIL_REGISTRATION_WITH_CONTRACT_WHEN_APPROVE_ACTIVITY = "SendMailRegistrationWithContractWhenApprove";
+    public static final String SEND_MAIL_REGISTRATION_REQUEST_ACTIVITY = "SendMailRegistrationRequest";
+    public static final String SEND_MAIL_REGISTRATION_APPROVE_ACTIVITY = "SendMailRegistrationApprove";
+    public static final String SEND_MAIL_ONBOARDING_APPROVE_ACTIVITY = "SendMailOnboardingApprove";
+    public static final String SEND_MAIL_CONFIRMATION_ACTIVITY = "SendMailConfirmation";
+
+
+    public static final String ONBOARDING_COMPLETION_ACTIVITY = "OnboardingCompletion";
+    public static final String CREATE_INSTITUTION_ACTIVITY = "CreateInstitution";
+    public static final String CREATE_ONBOARDING_ACTIVITY = "CreateOnboarding";
+    public static final String SEND_MAIL_COMPLETION_ACTIVITY = "SendMailCompletion";
+}
